refactor: menuData 상수로 분리
