+++ FE/src/components/header/gnb/menuBar/index.tsx
@@ -1,18 +1,9 @@
 import React from 'react';
+
 import * as Styled from 'components/header/gnb/menuBar/menubar.style';
+import menuData from 'constants/menuData';
 
 function MenuBar() {
-  type MenuType = {
-    id: number;
-    title: string;
-  };
-
-  const menuData: MenuType[] = [
-    { id: 1, title: '숙소' },
-    { id: 2, title: '체험' },
-    { id: 3, title: '온라인 체험' },
-  ];
-
   return (
     <Styled.MenuList>
       {menuData.map(({ id, title }) => (
