+++ FE/src/constants/menuData.ts
@@ -1,0 +1,12 @@
+type MenuType = {
+  id: number;
+  title: string;
+};
+
+const menuData: MenuType[] = [
+  { id: 1, title: '숙소' },
+  { id: 2, title: '체험' },
+  { id: 3, title: '온라인 체험' },
+];
+
+export default menuData;
